minor changes, switching to laptop
