+++ src/main/java/training/subsystems/Arm/Elbow/SimulationElbow/SimulationElbowConstants.java
@@ -1,4 +1,4 @@
-package training.subsystems.Arm.Elbow;
+package training.subsystems.Arm.Elbow.SimulationElbow;
 
 import edu.wpi.first.math.geometry.Rotation2d;
